+++ hooks/airtime_transferred_test.go
@@ -71,10 +71,10 @@ func TestAirtimeTransferred(t *testing.T) {
 
 	httpx.SetRequestor(httpx.NewMockRequestor(map[string][]httpx.MockResponse{
 		"https://airtime-api.dtone.com/cgi-bin/shop/topup": []httpx.MockResponse{
-			httpx.NewMockResponse(200, withCRLF(msisdnResponse)),
-			httpx.NewMockResponse(200, withCRLF(reserveResponse)),
-			httpx.NewMockResponse(200, withCRLF(topupResponse)),
-			httpx.NewMockResponse(200, "error_code=13\r\nerror_txt=Oops\r\n"),
+			httpx.NewMockResponse(200, withCRLF(msisdnResponse), nil),
+			httpx.NewMockResponse(200, withCRLF(reserveResponse), nil),
+			httpx.NewMockResponse(200, withCRLF(topupResponse), nil),
+			httpx.NewMockResponse(200, "error_code=13\r\nerror_txt=Oops\r\n", nil),
 		},
 	}))