+++ models/contacts_test.go
@@ -46,11 +46,19 @@ func TestElasticContacts(t *testing.T) {
 				"query":{
 					"bool":{
 						"must":[
-							{"match":{"name":{"query":"george"}}},
-							{"term":{"org_id":1}},
-							{"term":{"is_active":true}},
-							{"term":{"is_blocked":false}},
-							{"term":{"is_stopped":false}}
+							{ "bool":{
+								"must":[
+									{"term":{"org_id":1}},
+									{"term":{"is_active":true}},
+									{"match":{"name":{"query":"george"}}}
+								]
+							}},
+							{ "term":{
+								"is_blocked":false
+							}},
+							{"term":
+								{"is_stopped":false
+							}}
 						]
 					}
 				},
@@ -91,9 +99,13 @@ func TestElasticContacts(t *testing.T) {
 				"query":{
 					"bool":{
 						"must":[
-							{"match":{"name":{"query":"nobody"}}},
-							{"term":{"org_id":1}},
-							{"term":{"is_active":true}},
+							{"bool":
+								{"must":[
+									{"term":{"org_id":1}},
+									{"term":{"is_active":true}},
+									{"match":{"name":{"query":"nobody"}}}
+								]}
+							},
 							{"term":{"is_blocked":false}},
 							{"term":{"is_stopped":false}}
 						]
@@ -133,7 +145,7 @@ func TestElasticContacts(t *testing.T) {
 			assert.Error(t, err)
 		} else {
 			assert.NoError(t, err, "%d: error encountered performing query", i)
-			assert.JSONEq(t, tc.Request, es.LastBody, "%d: request mismatch", i)
+			assert.JSONEq(t, tc.Request, es.LastBody, "%d: request mismatch, got: %s", i, es.LastBody)
 			assert.Equal(t, tc.Contacts, ids, "%d: ids mismatch", i)
 		}
 	}